add caput log page
add using caputlog check for header navigation

add caputlog search layout component

add new line

add caputlogSearchConfig

require proptypes
