+++ src/components/header/Header.jsx
@@ -68,6 +68,14 @@ function Header() {
                     </ListItemIcon>
                     <ListItemText primary="Plotting" />
                   </ListItemButton>
+                  {import.meta.env.REACT_APP_USE_CAPUTLOG.toLowerCase() === "true" &&
+                    <ListItemButton key="CaputLog" component={NavLink} to="/caputlog" onClick={handleMenuToggle} divider color="primary">
+                      <ListItemIcon>
+                        <InfoIcon color="primary" />
+                      </ListItemIcon>
+                      <ListItemText primary="CaputLog" />
+                    </ListItemButton>
+                  }
                   <ListItemButton key="Services" component={NavLink} to="/services" onClick={handleMenuToggle} divider color="primary">
                     <ListItemIcon>
                       <InfoIcon color="primary" />
@@ -96,6 +104,11 @@ function Header() {
             {import.meta.env.REACT_APP_USE_AA.toLowerCase() === "true" &&
               <NavLink to="/plot" style={{ textDecoration: "none", color: 'inherit', paddingRight: 40 }}>
                 <Typography variant="h5">Plotting</Typography>
+              </NavLink>
+            }
+            {import.meta.env.REACT_APP_USE_CAPUTLOG.toLowerCase() === "true" &&
+              <NavLink to="/caputlog" style={{ textDecoration: "none", color: 'inherit', paddingRight: 40 }}>
+                <Typography variant="h5">Caput Log</Typography>
               </NavLink>
             }
             <NavLink to="/services" style={{ textDecoration: "none", color: 'inherit', paddingRight: 40 }}>
